feat: support react native new architecture (#42)
* feat: update change log

* feat: fix dependencies version static

* feat: update react native

* feat: update podfile

* feat: update MainApplication

* feat: update lockfiles

* feat: exclude

* feat: update dependencies

* feat: update dependencies

* feat: update depdendencies

* feat: update sdk version

* chore: bump version

* chore: bump version

* feat: update changelog

* feat: add beta publish step

* feat: support newarch, oldarch both

* feat: update interface

* feat: update node version

* feat: add eslint jest

* feat: change npm to yarn

* fix: fix lint

+++ android/src/oldarch/NotiflySdkSpec.kt
@@ -8,7 +8,7 @@ import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.bridge.Promise
 
 abstract class NotiflySdkSpec internal constructor(context: ReactApplicationContext) :
-  ReactContextBaseJavaModule(context) {
+        ReactContextBaseJavaModule(context) {
 
   abstract fun initialize(projectId: String, username: String, password: String, promise: Promise)
   abstract fun setUserId(userId: String?, promise: Promise)
@@ -18,12 +18,12 @@ abstract class NotiflySdkSpec internal constructor(context: ReactApplicationCont
   abstract fun setPhoneNumber(phoneNumber: String, promise: Promise)
   abstract fun setTimezone(timezone: String, promise: Promise)
   abstract fun trackEvent(
-    eventName: String,
-    params: ReadableMap?,
-    segmentationEventParamKeys: ReadableArray?,
-    promise: Promise,
+          eventName: String,
+          params: ReadableMap?,
+          segmentationEventParamKeys: ReadableArray?,
+          promise: Promise,
   )
-  abstract fun setLogLevel(logLevel: Int, promise: Promise)
+  abstract fun setLogLevel(logLevel: Double, promise: Promise)
   abstract fun disableInAppMessage(promise: Promise)
   abstract fun addNotificationClickListener(promise: Promise)
 }
